% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collidr.R
\docType{package}
\name{collidr-package}
\alias{collidr-package}
\title{collidr: Check for Namespace Collisions with Functions and Packages on CRAN}
\description{
collidr: Check for Namespace Collisions with Functions and Packages on CRAN
}
\details{
It has the goal of providing a convenient and native way of checking for namespace collisions when writing functions and packages.
}
\author{
Steve Condylios \email{steve.condylios@gmail.com}
}
