% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_function_collisions}
\alias{CRAN_function_collisions}
\title{Check for Namespace Collisions}
\usage{
CRAN_function_collisions(function_name, CRANdf)
}
\arguments{
\item{function_name}{A character string, or vector of character strings.}

\item{CRANdf}{Optionally provide an updated CRAN data.frame (obtain with getCRAN())}
}
\description{
Check for namespace collisions with functions on CRAN
}
\examples{

# Test single function name
function_name <- "a3.r2"
CRAN_function_collisions(function_name)

# Test multiple function names
function_name <- c("a3.r2", "xtable.A3")
CRAN_function_collisions(function_name)



}
