% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_package_collisions}
\alias{CRAN_package_collisions}
\title{Check for Namespace Collisions}
\usage{
CRAN_package_collisions(package_name, CRANdf)
}
\arguments{
\item{package_name}{A character string, or vector of character strings.}

\item{CRANdf}{Optionally provide an updated CRAN data.frame (obtain with getCRAN())}
}
\description{
Check for namespace collisions with packages on CRAN
}
\examples{


# Test single package name
package_name <- "dplyr"
CRAN_package_collisions(package_name)

# Test multiple package names
package_name <- c("dplyr", "data.frame", "gsubfn")
CRAN_package_collisions(package_name)



}
