% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_packages}
\alias{CRAN_packages}
\alias{CRANp}
\title{Show the packages that are on CRAN}
\usage{
CRAN_packages()

CRANp()
}
\description{
Retrieve a list of packages on CRAN.
}
\examples{

# Retrieve CRAN packages
CRAN_packages()


}
