% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANdf-data.R
\docType{data}
\name{CRANdf}
\alias{CRANdf}
\title{CRAN Package and Function Data: 1 May 2019}
\format{An object of class \code{data.frame} with 294190 rows and 2 columns.}
\source{
The Comprehensive R Archive Network, <https://cran.r-project.org/web/packages/available_packages_by_name.html>
}
\usage{
data(CRANdf)
}
\description{
Package and function data from The Comprehensive R Archive Network
}
\examples{
data(CRANdf)
packages <- CRANdf$package_names
functions <- CRANdf$function_names
}
\keyword{datasets}
