% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{getCRAN}
\alias{getCRAN}
\title{Retrieve a more up to date data.frame of packages and functions from CRAN}
\usage{
getCRAN(last_updated, api_key)
}
\description{
Retrieve a more up to date data.frame of packages and functions from CRAN
}
\examples{
\dontrun{
# Retrieve CRAN functions
CRAN_updated <- getCRAN()
}



}
