% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{getCRANpfd}
\alias{getCRANpfd}
\title{Retrieve a more up to date data.frame of packages and functions from CRAN}
\usage{
getCRANpfd(api_key)
}
\arguments{
\item{api_key}{An API key for collidr-api}
}
\description{
Retrieve a more up to date data.frame of packages and functions from CRAN
}
\examples{
\dontrun{
# Retrieve CRAN functions
updated_pfd <- getCRANpfd()
}



}
