% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages_and_functions_dataframe-data.R
\docType{data}
\name{packages_and_functions_dataframe}
\alias{packages_and_functions_dataframe}
\title{CRAN Package and Function Data: 1 May 2019}
\format{An object of class \code{data.frame} with 294190 rows and 2 columns.}
\source{
The Comprehensive R Archive Network, <https://cran.r-project.org/web/packages/available_packages_by_name.html>
}
\usage{
data(packages_and_functions_dataframe)
}
\description{
Package and function data from The Comprehensive R Archive Network
}
\examples{
data(packages_and_functions_dataframe)
packages <- packages_and_functions_dataframe$package_names
functions <- packages_and_functions_dataframe$function_names
}
\keyword{datasets}
