% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{retrieve_paf_dataframe}
\alias{retrieve_paf_dataframe}
\title{Retrieve a more up to date data.frame of packages and functions from CRAN}
\usage{
retrieve_paf_dataframe()
}
\description{
Retrieve a more up to date data.frame of packages and functions from CRAN
}
\examples{
\dontrun{
# Retrieve CRAN functions
new_packages_and_functions_data <- retrieve_paf_dataframe()
}


}
