% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_functions}
\alias{CRAN_functions}
\title{Show the functions that are on CRAN}
\usage{
CRAN_functions()
}
\description{
Retrieve functions from CRAN.
}
\examples{

# Retrieve CRAN functions
CRAN_functions()


}
