% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_functions}
\alias{CRAN_functions}
\alias{CRANf}
\title{Show the functions that are on CRAN}
\usage{
CRAN_functions(CRANdf)

CRANf(CRANdf)
}
\arguments{
\item{CRANdf}{Optionally provide an updated CRAN data.frame (obtain with getCRAN())}
}
\description{
Show functions from CRAN.
}
\examples{

# Show CRAN functions
CRAN_functions()


}
