% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_packages}
\alias{CRAN_packages}
\alias{CRANp}
\title{Show the packages that are on CRAN}
\usage{
CRAN_packages(CRANdf)

CRANp(CRANdf)
}
\arguments{
\item{CRANdf}{Optionally provide an updated CRAN data.frame (obtain with getCRAN())}
}
\description{
Show a list of packages on CRAN.
}
\examples{

# Show CRAN packages
CRAN_packages()


}
