% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_packages_and_functions}
\alias{CRAN_packages_and_functions}
\title{Show the packages and functions that are on CRAN}
\usage{
CRAN_packages_and_functions()
}
\description{
Retrieve functions from CRAN.
}
\examples{

# Retrieve CRAN packages and functions
CRAN_packages_and_functions()


}
