% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{CRAN_packages_and_functions}
\alias{CRAN_packages_and_functions}
\alias{CRANpf}
\title{Show the packages and functions that are on CRAN}
\usage{
CRAN_packages_and_functions(CRANdf)

CRANpf(CRANdf)
}
\arguments{
\item{CRANdf}{Optionally provide an updated CRAN data.frame (obtain with getCRAN())}
}
\description{
Show functions from CRAN.
}
\examples{

# Show CRAN packages and functions
CRAN_packages_and_functions()


}
