% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{getCRAN}
\alias{getCRAN}
\title{Retrieve a more up to date data.frame of packages and functions from CRAN}
\usage{
getCRAN(last_updated, api_key)
}
\arguments{
\item{last_updated}{Set to TRUE to return the timestamp of the last update to the CRAN
database file. Access via attributes()}

\item{api_key}{An API key for collidr-api}
}
\description{
Retrieve a more up to date data.frame of packages and functions from CRAN
}
\examples{
\dontrun{
# Retrieve CRAN functions
CRAN_updated <- getCRAN()
}



}
